pyproject.toml
[project]
name = "lastversion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imaplib2>=3.6",
    "json5>=0.12.0",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.28",
    "openai>=1.97.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "pymongo>=4.13.2",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "reportlab>=4.4.3",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.47.0",
    "streamlit-aggrid>=1.1.7",
    "streamlit-autorefresh>=1.0.1",
    "streamlit-pdf-viewer>=0.0.26",
]
